Authen Staff and get avatar staff + some adjust for better fault tolerance

+++ Backend/TestForASPWebAPI/TestForASPWebAPI/Controllers/CustomerController.cs
@@ -44,7 +44,7 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // GET api/<ValuesController>/5
-        [HttpGet("GetCustomerById")]
+        [HttpGet("GetCustomerById/{id}")]
         public async Task<IActionResult> Get(int id)
         {
             DBController dbController = DBController.GetInstance();
@@ -89,15 +89,18 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // DELETE api/<ValuesController>/5
-        [HttpDelete("Delete")]
-        public void Delete(int id)
+        [HttpDelete("Delete/{id}")]
+        public async Task<IActionResult> Delete(int id)
         {
+            if (!await CustomerExists(id)) { return NotFound("Customer not found!"); }
+
             string command = $"DELETE FROM Customer WHERE Id = {id}";
             DBController dbController = DBController.GetInstance();
             dbController.DeleteData(command);
-            return;
+
+            return NoContent();
         }
-        [HttpGet("Exists")]
+        [HttpGet("Exists/{id}")]
         public async Task<bool> CustomerExists(int id)
         {
             string command = $"SELECT * FROM Customer WHERE Id = {id}";
