+++ Backend/TestForASPWebAPI/TestForASPWebAPI/Controllers/PriceController.cs
@@ -45,7 +45,7 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // GET api/<ValuesController>/5
-        [HttpGet("GetPriceById")]
+        [HttpGet("GetPriceById/{id}")]
         public async Task<IActionResult> Get(int id)
         {
             DBController dbController = DBController.GetInstance();
@@ -91,15 +91,18 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // DELETE api/<ValuesController>/5
-        [HttpDelete("Delete")]
-        public void Delete(int id)
+        [HttpDelete("Delete/{id}")]
+        public async Task<IActionResult> Delete(int id)
         {
+            if (!await PriceExists(id)) { return NotFound("Price not found!"); }
+
             string command = $"DELETE FROM Price WHERE Id = {id}";
             DBController dbController = DBController.GetInstance();
             dbController.DeleteData(command);
-            return;
+
+            return NoContent();
         }
-        [HttpGet("Exists")]
+        [HttpGet("Exists/{id}")]
         public async Task<bool> PriceExists(int id)
         {
             string command = $"SELECT * FROM Price WHERE Id = {id}";
