+++ Backend/TestForASPWebAPI/TestForASPWebAPI/Controllers/ProductInstanceController.cs
@@ -44,7 +44,7 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // GET api/<ValuesController>/5
-        [HttpGet("GetProductInstanceById")]
+        [HttpGet("GetProductInstanceById/{id}")]
         public async Task<IActionResult> Get(int id)
         {
             DBController dbController = DBController.GetInstance();
@@ -89,15 +89,18 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // DELETE api/<ValuesController>/5
-        [HttpDelete("Delete")]
-        public void Delete(int id)
+        [HttpDelete("Delete/{id}")]
+        public async Task<IActionResult> Delete(int id)
         {
+            if (!await ProductInstanceExists(id)) { return NotFound("Instance not found!"); }
+
             string command = $"DELETE FROM ProductInstance WHERE Id = {id}";
             DBController dbController = DBController.GetInstance();
             dbController.DeleteData(command);
-            return;
+
+            return NoContent();
         }
-        [HttpGet("Exists")]
+        [HttpGet("Exists/{id}")]
         public async Task<bool> ProductInstanceExists(int id)
         {
             string command = $"SELECT * FROM ProductInstance WHERE Id = {id}";
