+++ Backend/TestForASPWebAPI/TestForASPWebAPI/Controllers/ProductLineController.cs
@@ -46,7 +46,7 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // GET api/<ValuesController>/5
-        [HttpGet("GetProductLineById")]
+        [HttpGet("GetProductLineById/{id}")]
         public async Task<IActionResult> Get(int id)
         {
             DBController dbController = DBController.GetInstance();
@@ -93,15 +93,18 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // DELETE api/<ValuesController>/5
-        [HttpDelete("Delete")]
-        public void Delete(int id)
+        [HttpDelete("Delete/{id}")]
+        public async Task<IActionResult> Delete(int id)
         {
+            if (!await ProductLineExists(id)) { return NotFound("Product Line not found!"); }
+
             string command = $"DELETE FROM ProductLine WHERE Id = {id}";
             DBController dbController = DBController.GetInstance();
             dbController.DeleteData(command);
-            return;
+
+            return NoContent();
         }
-        [HttpGet("Exists")]
+        [HttpGet("Exists/{id}")]
         public async Task<bool> ProductLineExists(int id)
         {
             string command = $"SELECT * FROM ProductLine WHERE Id = {id}";
