+++ Backend/TestForASPWebAPI/TestForASPWebAPI/Controllers/ProductSpecificationController.cs
@@ -42,7 +42,7 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // GET api/<ValuesController>/5
-        [HttpGet("GetProductSpecificationById")]
+        [HttpGet("GetProductSpecificationById/{id}")]
         public async Task<IActionResult> Get(int id)
         {
             DBController dbController = DBController.GetInstance();
@@ -85,15 +85,18 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // DELETE api/<ValuesController>/5
-        [HttpDelete("Delete")]
-        public void Delete(int id)
+        [HttpDelete("Delete/{id}")]
+        public async Task<IActionResult> Delete(int id)
         {
+            if (!await ProductSpecificationExists(id)) { return NotFound("Specification not found!"); }
+
             string command = $"DELETE FROM ProductSpecification WHERE Id = {id}";
             DBController dbController = DBController.GetInstance();
             dbController.DeleteData(command);
-            return;
+
+            return NoContent();
         }
-        [HttpGet("Exists")]
+        [HttpGet("Exists/{id}")]
         public async Task<bool> ProductSpecificationExists(int id)
         {
             string command = $"SELECT * FROM ProductSpecification WHERE Id = {id}";
