+++ Backend/TestForASPWebAPI/TestForASPWebAPI/Controllers/PromotionController.cs
@@ -48,7 +48,7 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // GET api/<ValuesController>/5
-        [HttpGet("GetPromotionById")]
+        [HttpGet("GetPromotionById/{id}")]
         public async Task<IActionResult> Get(int id)
         {
             DBController dbController = DBController.GetInstance();
@@ -97,15 +97,18 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // DELETE api/<ValuesController>/5
-        [HttpDelete("Delete")]
-        public void Delete(int id)
+        [HttpDelete("Delete/{id}")]
+        public async Task<IActionResult> Delete(int id)
         {
+            if (!await PromotionExists(id)) { return NotFound("Promotion not found!"); }
+
             string command = $"DELETE FROM Promotion WHERE Id = {id}";
             DBController dbController = DBController.GetInstance();
             dbController.DeleteData(command);
-            return;
+
+            return NoContent();
         }
-        [HttpGet("Exists")]
+        [HttpGet("Exists/{id}")]
         public async Task<bool> PromotionExists(int id)
         {
             string command = $"SELECT * FROM Promotion WHERE Id = {id}";
