+++ Backend/TestForASPWebAPI/TestForASPWebAPI/Controllers/PromotionUsageController.cs
@@ -42,7 +42,7 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // GET api/<ValuesController>/5
-        [HttpGet("GetPromotionUsageById")]
+        [HttpGet("GetPromotionUsageById/{id}")]
         public async Task<IActionResult> Get(int id)
         {
             DBController dbController = DBController.GetInstance();
@@ -85,15 +85,18 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // DELETE api/<ValuesController>/5
-        [HttpDelete("Delete")]
-        public void Delete(int id)
+        [HttpDelete("Delete/{id}")]
+        public async Task<IActionResult> Delete(int id)
         {
+            if (!await PromotionUsageExists(id)) { return NotFound("Promo Usage not found!"); }
+
             string command = $"DELETE FROM PromotionUsage WHERE Id = {id}";
             DBController dbController = DBController.GetInstance();
             dbController.DeleteData(command);
-            return;
+
+            return NoContent();
         }
-        [HttpGet("Exists")]
+        [HttpGet("Exists/{id}")]
         public async Task<bool> PromotionUsageExists(int id)
         {
             string command = $"SELECT * FROM PromotionUsage WHERE Id = {id}";
