+++ Backend/TestForASPWebAPI/TestForASPWebAPI/Controllers/SpecificationTypeController.cs
@@ -41,7 +41,7 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // GET api/<ValuesController>/5
-        [HttpGet("GetSpecificationTypeById")]
+        [HttpGet("GetSpecificationTypeById/{id}")]
         public async Task<IActionResult> Get(int id)
         {
             DBController dbController = DBController.GetInstance();
@@ -83,15 +83,18 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // DELETE api/<ValuesController>/5
-        [HttpDelete("Delete")]
-        public void Delete(int id)
+        [HttpDelete("Delete/{id}")]
+        public async Task<IActionResult> Delete(int id)
         {
+            if (!await SpecificationTypeExists(id)) { return NotFound("Type not found!"); }
+
             string command = $"DELETE FROM SpecificationType WHERE Id = {id}";
             DBController dbController = DBController.GetInstance();
             dbController.DeleteData(command);
-            return;
+
+            return NoContent();
         }
-        [HttpGet("Exists")]
+        [HttpGet("Exists/{id}")]
         public async Task<bool> SpecificationTypeExists(int id)
         {
             string command = $"SELECT * FROM SpecificationType WHERE Id = {id}";
